monelanal/requirements.txt
flask>=3.0
requests>=2.28
sqlalchemy>=2.0
python-dotenv>=1.0
tzdata>=2024.1
